# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AgentWorkspaceFileKind = typeof AgentWorkspaceFileKind[keyof typeof AgentWorkspaceFileKind];


export const AgentWorkspaceFileKind = {
  file: 'file',
  folder: 'folder',
} as const;

export interface AgentWorkspaceFile {
  path: string;
  kind: AgentWorkspaceFileKind;
  /** @nullable */
  language: string | null;
  size: number;
}

export type AgentWorkspaceContents = {[key: string]: string};

export interface AgentWorkspace {
  projectName: string;
  branch: string;
  status: string;
  files: AgentWorkspaceFile[];
  contents: AgentWorkspaceContents;
}

export interface AgentTaskInput {
  /** @minLength 1 */
  prompt: string;
}

export interface AgentFileUpdateInput {
  /** @minLength 1 */
  path: string;
  content: string;
}

export interface AgentChange {
  path: string;
  summary: string;
}

export interface AgentTaskResult {
  message: string;
  changes: AgentChange[];
  commands: string[];
  output: string;
  workspace: AgentWorkspace;
}