# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: agent
    description: Coding workspace and agent operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /agent/workspace:
    get:
      operationId: getAgentWorkspace
      tags: [agent]
      summary: Read the active coding workspace
      responses:
        "200":
          description: Workspace files and metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentWorkspace"
        "500":
          description: Workspace unavailable
  /agent/run:
    post:
      operationId: runAgentTask
      tags: [agent]
      summary: Ask the coding agent to modify the workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentTaskInput"
      responses:
        "200":
          description: Agent result with applied changes and command output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentTaskResult"
        "400":
          description: Invalid task
        "503":
          description: Agent provider is not configured
  /agent/file:
    put:
      operationId: updateAgentFile
      tags: [agent]
      summary: Save a file in the active coding workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentFileUpdateInput"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentWorkspace"
        "400":
          description: Invalid file path or content
  /agent/preview:
    get:
      operationId: getAgentPreview
      tags: [agent]
      summary: Serve the active workspace preview
      responses:
        "200":
          description: Workspace preview HTML
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Preview entrypoint unavailable
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AgentWorkspaceFile:
      type: object
      properties:
        path:
          type: string
        kind:
          type: string
          enum: [file, folder]
        language:
          type: ["string", "null"]
        size:
          type: integer
      required: [path, kind, language, size]
    AgentWorkspace:
      type: object
      properties:
        projectName:
          type: string
        branch:
          type: string
        status:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/AgentWorkspaceFile"
        contents:
          type: object
          additionalProperties:
            type: string
      required: [projectName, branch, status, files, contents]
    AgentTaskInput:
      type: object
      properties:
        prompt:
          type: string
          minLength: 1
      required: [prompt]
    AgentFileUpdateInput:
      type: object
      properties:
        path:
          type: string
          minLength: 1
        content:
          type: string
      required: [path, content]
    AgentChange:
      type: object
      properties:
        path:
          type: string
        summary:
          type: string
      required: [path, summary]
    AgentTaskResult:
      type: object
      properties:
        message:
          type: string
        changes:
          type: array
          items:
            $ref: "#/components/schemas/AgentChange"
        commands:
          type: array
          items:
            type: string
        output:
          type: string
        workspace:
          $ref: "#/components/schemas/AgentWorkspace"
      required: [message, changes, commands, output, workspace]