# lib/api-zod/src/generated/types/agentFileUpdateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AgentFileUpdateInput {
  /** @minLength 1 */
  path: string;
  content: string;
}